+++ gluecodium/src/test/resources/smoke/generic_types/output/cbridge/src/GenericCollections.cpp
@@ -8,12 +8,14 @@
 #include "cbridge/include/smoke/cbridge_DummyClass.h"
 #include "cbridge/include/smoke/cbridge_DummyInterface.h"
 #include "cbridge/include/smoke/cbridge_GenericTypesWithCompoundTypes.h"
+#include "cbridge/include/smoke/cbridge_YetAnotherDummyClass.h"
 #include "cbridge_internal/include/BaseHandleImpl.h"
 #include "gluecodium/Optional.h"
 #include "smoke/AnotherDummyClass.h"
 #include "smoke/DummyClass.h"
 #include "smoke/DummyInterface.h"
 #include "smoke/GenericTypesWithCompoundTypes.h"
+#include "smoke/YetAnotherDummyClass.h"
 #include <memory>
 #include <new>
 #include <string>
@@ -402,6 +404,33 @@ void foobar_ArrayOf_smoke_GenericTypesWithCompoundTypes_SomeEnum_release_optiona
 }
 _baseRef foobar_ArrayOf_smoke_GenericTypesWithCompoundTypes_SomeEnum_unwrap_optional_handle(_baseRef handle) {
     return reinterpret_cast<_baseRef>( &**reinterpret_cast<::gluecodium::optional<std::vector<::smoke::GenericTypesWithCompoundTypes::SomeEnum>>*>( handle ) );
+}
+_baseRef foobar_ArrayOf_smoke_YetAnotherDummyClass_create_handle() {
+    return reinterpret_cast<_baseRef>( new std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>( ) );
+}
+_baseRef foobar_ArrayOf_smoke_YetAnotherDummyClass_copy_handle(_baseRef handle) {
+    return reinterpret_cast<_baseRef>( new std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>( *reinterpret_cast<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>*>( handle ) ) );
+}
+void foobar_ArrayOf_smoke_YetAnotherDummyClass_release_handle(_baseRef handle) {
+    delete reinterpret_cast<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>*>( handle );
+}
+uint64_t foobar_ArrayOf_smoke_YetAnotherDummyClass_count(_baseRef handle) {
+    return Conversion<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>>::toCpp( handle ).size( );
+}
+_baseRef foobar_ArrayOf_smoke_YetAnotherDummyClass_get( _baseRef handle, uint64_t index ) { return Conversion<std::shared_ptr<::smoke::YetAnotherDummyClass>>::referenceBaseRef(Conversion<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>>::toCpp( handle )[index]);
+}
+void foobar_ArrayOf_smoke_YetAnotherDummyClass_append( _baseRef handle, _baseRef item )
+{
+Conversion<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>>::toCpp(handle).push_back(Conversion<std::shared_ptr<::smoke::YetAnotherDummyClass>>::toCpp(item));
+}
+_baseRef foobar_ArrayOf_smoke_YetAnotherDummyClass_create_optional_handle() {
+    return reinterpret_cast<_baseRef>( new ( std::nothrow ) ::gluecodium::optional<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>>( std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>( ) ) );
+}
+void foobar_ArrayOf_smoke_YetAnotherDummyClass_release_optional_handle(_baseRef handle) {
+    delete reinterpret_cast<::gluecodium::optional<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>>*>( handle );
+}
+_baseRef foobar_ArrayOf_smoke_YetAnotherDummyClass_unwrap_optional_handle(_baseRef handle) {
+    return reinterpret_cast<_baseRef>( &**reinterpret_cast<::gluecodium::optional<std::vector<std::shared_ptr<::smoke::YetAnotherDummyClass>>>*>( handle ) );
 }
 _baseRef foobar_MapOf__Float_To__Double_create_handle() {
     return reinterpret_cast<_baseRef>( new ( std::nothrow ) std::unordered_map<float, double >() );